refactor: mypage関連コンポーネントのディレクトリ作成
