+++ src/components/organisms/mypage/BasicInformation/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import BasicInformationTable from "../BasicInformationTable";
-import IconImage from "../../atoms/IconImage";
-import { UserInfo } from "../../../types";
+import IconImage from "../IconImage";
+import { UserInfo } from "../../../../types";
 import styles from "./style.module.css";
 
 type BasicInformationProps = {
@@ -11,7 +11,7 @@ type BasicInformationProps = {
 const BasicInformation = ({ userInfo }: BasicInformationProps) => {
   return (
     <div className={styles.container}>
-      <IconImage src={userInfo.icon_url}/>
+      <IconImage src={userInfo.icon_url} />
       <BasicInformationTable userInfo={userInfo} />
     </div>
   );
